requirements.txt
Flask
Werkzeug
pandas
numpy
requests
beautifulsoup4
python-dotenv
lxml
APScheduler
cachetools
mysql-connector-python
bcrypt
PyJWT
pytz
flask-cors
webauthn
yt-dlp
gunicorn
pysocks
cryptography